ScreenSync_v2/requirements.txt
tinytuya>=1.2.0
Pillow>=8.0.0
paho-mqtt
#tkinter  # This is usually included with Python's standard library
matplotlib
mss
platformdirs
flux_led
python-kasa>=0.10.2